+++ src/main/open_url.ts
@@ -1,5 +1,7 @@
 import { app as rawApp, ipcMain } from 'electron'
 import type { EventEmitter } from 'events'
+import { readFile } from 'fs/promises'
+import { basename } from 'path'
 import { getLogger } from '../shared/logger'
 import { supportedURISchemes } from './application-constants'
 import { showDeltaChat } from './tray'
@@ -65,10 +67,51 @@ app.on('open-url', (event, url) => {
   open_url(url)
 })
 
-// Iterate over arguments and look out for uris
-export function openUrlFromArgv(argv: string[]) {
+async function handleWebxdcFileOpen(path: string) {
+  log.info('open file', path)
+  if (!path.endsWith('.xdc')) {
+    log.info('handleWebxdcFileOpen, path does not contain .xdc', path)
+    return
+  }
+  app.focus()
+  window?.focus()
+
+  // hacky code - abuses webxdc sendToChat
+  // todo make this code nicer and maybe show even a custom dialog that shows what is being sent?
+  const buffer = await readFile(path)
+  if (!app.ipcReady) {
+    await new Promise(res => (app as any).once('ipcReady', res))
+  }
+  if (!frontend_ready) {
+    await new Promise(res => ipcMain.once('frontendReady', res))
+  }
+  window?.webContents.send(
+    'webxdc.sendToChat',
+    { file_name: basename(path), file_content: buffer.toString('base64') },
+    null
+  )
+}
+
+app.on('open-file', async (event, path) => {
+  if (event) {
+    event.preventDefault()
+  }
+  handleWebxdcFileOpen(path)
+})
+
+// Iterate over arguments and look out for uris and webxdc file paths
+export function openUrlsAndFilesFromArgv(argv: string[]) {
   args_loop: for (let i = 1; i < argv.length; i++) {
     const arg = argv[i]
+
+    if (arg.endsWith('.xdc')) {
+      log.debug(
+        'open-url: process something that looks like it could be a webxc file:',
+        arg
+      )
+      handleWebxdcFileOpen(arg)
+      continue
+    }
 
     if (!arg.includes(':')) {
       continue
@@ -93,7 +136,8 @@ export function openUrlFromArgv(argv: string[]) {
 
 app.on('second-instance', (_event, argv) => {
   log.debug('Someone tried to run a second instance')
-  openUrlFromArgv(argv)
+  openUrlsAndFilesFromArgv(argv)
+  // open file from argv
   if (window) {
     showDeltaChat()
   }